Prepare release 4.4

See merge request nonsmooth/siconos!14

+++ kernel/src/modelingTools/DynamicalSystem.cpp
@@ -25,7 +25,7 @@
 #include <iostream>
 
 
-unsigned int DynamicalSystem::__count = 0;
+size_t DynamicalSystem::__count = 0;
 
 void DynamicalSystem::_init()
 {
